Making waitForHydration re-usable
