+++ app/store/utils.js
@@ -40,3 +40,25 @@ export function transformSet(incoming, setTransforms, toStorage = true) {
 
     return state;
 }
+
+export function waitForHydration(store, callback) {
+    let unsubscribeFromStore;
+    return new Promise((resolve) => {
+        if (__DEV__) {
+            // when in DEV mode resetting the app can get into a white screen
+            resolve();
+            return;
+        }
+        const subscription = () => {
+            if (store.getState().views.root.hydrationComplete) {
+                unsubscribeFromStore();
+                if (callback && typeof callback === 'function') {
+                    callback();
+                }
+                resolve();
+            }
+        };
+
+        unsubscribeFromStore = store.subscribe(subscription);
+    });
+}
