+++ app/utils/push_notifications.js
@@ -23,6 +23,7 @@ import {getCurrentServerUrl, getAppCredentials} from 'app/init/credentials';
 import PushNotifications from 'app/push_notifications';
 import {getCurrentLocale} from 'app/selectors/i18n';
 import EphemeralStore from 'app/store/ephemeral_store';
+import {waitForHydration} from 'app/store/utils';
 import {t} from 'app/utils/i18n';
 
 class PushNotificationUtils {
@@ -64,9 +65,6 @@ class PushNotificationUtils {
 
     onPushNotification = async (deviceNotification) => {
         const {dispatch, getState} = this.store;
-        let unsubscribeFromStore = null;
-        let stopLoadingNotification = false;
-
         const {data, foreground, message, userInteraction} = deviceNotification;
         const notification = {
             data,
@@ -88,15 +86,9 @@ class PushNotificationUtils {
                         this.loadFromNotification(notification);
                     }, 0);
                 } else {
-                    const waitForHydration = () => {
-                        if (getState().views.root.hydrationComplete && !stopLoadingNotification) {
-                            stopLoadingNotification = true;
-                            unsubscribeFromStore();
-                            this.loadFromNotification(notification);
-                        }
-                    };
-
-                    unsubscribeFromStore = this.store.subscribe(waitForHydration);
+                    waitForHydration(this.store, () => {
+                        this.loadFromNotification(notification);
+                    });
                 }
             }
         }
@@ -164,8 +156,7 @@ class PushNotificationUtils {
     };
 
     onRegisterDevice = (data) => {
-        const {dispatch, getState} = this.store;
-        let unsubscribeFromStore = null;
+        const {dispatch} = this.store;
 
         let prefix;
         if (Platform.OS === 'ios') {
@@ -180,15 +171,10 @@ class PushNotificationUtils {
         EphemeralStore.deviceToken = `${prefix}:${data.token}`;
 
         // TODO: Remove when realm is ready
-        const waitForHydration = () => {
-            if (getState().views.root.hydrationComplete && !this.configured) {
-                this.configured = true;
-                dispatch(setDeviceToken(EphemeralStore.deviceToken));
-                unsubscribeFromStore();
-            }
-        };
-
-        unsubscribeFromStore = this.store.subscribe(waitForHydration);
+        waitForHydration(this.store, () => {
+            this.configured = true;
+            dispatch(setDeviceToken(EphemeralStore.deviceToken));
+        });
     };
 
     getNotification = () => {
